requirements.txt
# Install WikiApiary requirements easily with
# sudo pip install -r requirements.txt
# sudo pip install --upgrade -r requirements.txt
#
# External requirements are mysqld and redis-server

# We need to connect to mysql
mysql-python
yoyo-migrations

# Use requests to make it easier to do things with URLs
requests

# This is the library the bots use to talk to wikiapiary.com
simplemediawiki
kitchen

# We use this to get data from MaxMind and store information
# into wikiapiary.com from MaxMind database
pygeoip

# This is used during audit process to find things like the
# logo for the wiki
beautifulsoup4

# Parse JSON easier
simplejson

# Coveralls for code coverage
python-coveralls

# We use Celery to manage our queue, with redis as the backend
redis
celery[redis]